Fix problems with overflowing boxes using flex (BL-10053)
bloom-translationGroups that are vertically centered or aligned bottom use display:flex. When they overflow, the browser tries to compress the text, sometimes to the point of overlapping and being unreadable. If we use flex-shrink:0 to prevent this, nicescroll fails to provide scrolling.  This PR simply turns this off for overflowing blocks.