fix: prevent a possible render problem
I simplified this logic drastically. If the old logic has a purpose, it should be added to preprocessUrl() with comments. As far as I can see, state.bookUrl always has a leading slash, and in that case the logic I've removed never assigns this.sourceUrl to be anything other than newSourceUrl.